LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/wheel-event-transactions-basic_include=scroll-over-scrollable-child-expected.txt


FAIL Wheel events should be captured to one target #2 assert_equals: Wheel event at index `0` does not have the expected target expected Element node <div id="secondRootSpacer" class="spacer" style="backgrou... but got Element node <div id="firstInnerSpacer" class="spacer" style="backgrou...
